requirements.txt
RPi.GPIO
astral
telepot
urllib3